Put MuxedAccount back into account.js to simplify the diff
